manage optionnal output for ericscript

+++ main.nf
@@ -509,8 +509,13 @@ process ericscript {
         -o ./tmp \\
         ${reads}
 
-    mv tmp/fusions.results.filtered.tsv ${sample}_ericscript.tsv
-    mv tmp/fusions.results.total.tsv ${sample}_ericscript_total.tsv
+    if [[ -f "tmp/fusions.results.filtered.tsv" ]]; then
+        mv tmp/fusions.results.filtered.tsv ${sample}_ericscript.tsv
+    fi 
+
+    if [[ -f "tmp/fusions.results.filtered.tsv" ]]; then
+        mv tmp/fusions.results.total.tsv ${sample}_ericscript_total.tsv
+    fi 
     """
 }
